requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
scipy
faiss-cpu
